Expression enum → protocols + structs
This has a few benefits:

1. Expressions can constrain certain members to be a specific type of
   expression. For example, `FunctionExpression`’s `protocol` can now be
   statically guaranteed to be a `ProtocolExpression`.

2. We can differentiate between value expressions and top-level expressions.
   This can be useful, for example, in a REPL where value expressions are
   valid at the top level vs. file compilation where only true top-level
   expressions are.

This adds some grossness with equality checking, however. Since we want to be
able to have `[Expression]` we can’t make that protocol have a self-requirement.
So we have to do the fun dance of adding `equals(other: Expression) -> Bool` to
the `Expression` protocol (see WWDC 2015 Session 408).